+++ people-pro/package.json
@@ -10,6 +10,7 @@
     "start": "npm run serve"
   },
   "dependencies": {
+    "@popperjs/core": "^2.11.8",
     "bootstrap": "^5.3.2",
     "core-js": "^3.8.3",
     "v-mask": "^2.3.0",
