+++ people-pro/src/router.js
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import CadastroColaborador from './colaboradores/CadastroColaborador.vue'; // Importe o componente corretamente
+import CadastroColaborador from './pages/colaboradores/Cadastro';
 
 const router = createRouter({
   history: createWebHistory(),
@@ -9,7 +9,6 @@ const router = createRouter({
       name: 'CadastroColaborador',
       component: CadastroColaborador,
     },
-    // Outras rotas aqui, se necessário
   ],
 });
