+++ people-pro/src/style.css
@@ -39,6 +39,7 @@
   color: #fff;
   cursor: pointer;
   border-radius: 5px;
+  margin: 0 2% 2% 0;
 }
 
 .people-pro-form-btn:hover {
